Change SourceMap location tracking and getParents method.
Use getStart() for ts.Nodes instead of their pos to get their actual location in the file and not the position right after the previous token.
Use getChildren().forEach() instead of forEachChild() to truly get all AST nodes (and not skip ones that are basically just enclosures, like method and function definitions).
Rename getParent() to getEnclosingNodes() and return all ts.Nodes that intersect with the given SourcePart's [pos, end]