Move metric collection to separate classes. Add accessor functions to SourceMap.
Add IMetricCollector interface to have a common starting point for all metric collection algorithms that can be called while the source map is iterating the AST anyways.
Add accessor to get comments that belong to a node following a stricter notion of the general TS style:
Comments leading before a line belong to the line, but only the ones that are directly connected (i.e., no empty line) to the node.
Trailing comments in the same line do belong to nodes in this line.