Several smaller complexity calculation and evaluation changes.
Use multiple lines of comments as section division, as it creates quite a huge vertical space.
Don't save section complexities with value 0 (which would be created between multiple empty lines).
Only mark complex lines in a section to require a comment if the section start is already commented instead of just using "the most complex one" (which could still have a low complexity).
Don't add the distance to the section start to the line complexity evaluation.