Add tests for comment classifier and fix behavior to match tests.
Move createSourceMap function to testHelper.
Don't add additional member or header classifications if a comment has been classified as inline.
Store a flag in a comment to reflect whether it is trailing or not.
No longer use the first node in the commented line to determine comment class, but the most enclosing one to ignore leading semicolons etc.