Adjust comment quality evaluation.
Lower the comment quality if more than 40% of the words of the comment name are in the declaration name.
Include parameter names in declaration names if evaluating non-JSDoc-comments.
No longer include parameter types in JSDoc parameter comment evaluation.
Adjust the comment quality in both directions based on good and bad parameter comments:
	Adjust negatively for each not commented parameter.
	Adjust positively if the parameter comment quality is more positive than negative.
	Adjust negatively if the parameter comment quality is more negative than positive.